refactor(FieldChange): move styled component styles from the component file to the styled file
