+++ packages/@sanity/field/src/diff/components/FieldChange.tsx
@@ -1,8 +1,7 @@
 import {useDocumentOperation} from '@sanity/react-hooks'
 import React, {useCallback, useContext, useState} from 'react'
 import {unstable_useCheckDocumentPermission as useCheckDocumentPermission} from '@sanity/base/hooks'
-import {rem, Stack, Box, Button, Text, Grid, Popover, useClickOutside} from '@sanity/ui'
-import styled from 'styled-components'
+import {Stack, Box, Button, Text, Grid, Popover, useClickOutside} from '@sanity/ui'
 import {undoChange} from '../changes/undoChange'
 import {DiffContext} from '../contexts/DiffContext'
 import {FieldChangeNode, OperationsAPI} from '../../types'
@@ -14,39 +13,7 @@ import {FallbackDiff} from './FallbackDiff'
 import {RevertChangesButton} from './RevertChangesButton'
 import {ValueError} from './ValueError'
 
-import {BoxContentWrapper} from './FieldChange.styled'
-
-const FieldChangeContainer = styled.div`
-  --field-change-error: ${({theme}) => theme.sanity.color.solid.critical.enabled.bg};
-  &[data-revert-all-changes-hover] [data-revert-all-hover]::before {
-    border-left: 2px solid var(--field-change-error);
-  }
-`
-
-const DiffBorder = styled.div`
-  --field-change-error: ${({theme}) => theme.sanity.color.solid.critical.enabled.bg};
-  --diff-inspect-padding-xsmall: ${({theme}) => rem(theme.sanity.space[1])};
-  --diff-inspect-padding-small: ${({theme}) => rem(theme.sanity.space[2])};
-
-  position: relative;
-  padding: var(--diff-inspect-padding-xsmall) 0 var(--diff-inspect-padding-xsmall)
-    var(--diff-inspect-padding-small);
-
-  &::before {
-    content: '';
-    display: block;
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    border-left: 1px solid var(--card-border-color);
-  }
-
-  &[data-error]:hover::before,
-  &[data-revert-field-hover]:hover::before {
-    border-left: 2px solid var(--field-change-error);
-  }
-`
+import {FieldChangeContainer, DiffBorder, BoxContentWrapper} from './FieldChange.styled'
 
 export function FieldChange({
   change,
